added loading state
